src/requirements.txt
asgiref==3.7.2
Django==4.2.7
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
django-sesame==3.2.1
websockets==12.0